finalstep

+++ src/main/java/web/service/UserService.java
@@ -6,7 +6,7 @@ import java.util.List;
 public interface UserService {
     void addUser(User user);
     User getUser(long id);
-    void deleteUser(int id);
+    void deleteUser(long id);
     void updateUser(User user);
     List<User> getAllUsers();
 }
